+++ CHANGES.md
@@ -17,6 +17,7 @@ Contributors to this version: Joseph Siddons (@josidd), Steven Chan (@stchan), R
 
 ### Breaking changes
 
+* Removed `variogram.LinearVariagram` and `variogram.PowerVariogram` (!69)
 * `perturbation` module is renamed to `stochastic` (!65)
 * All job-specific files are removed from the library (!56)
 
@@ -24,6 +25,7 @@ Contributors to this version: Joseph Siddons (@josidd), Steven Chan (@stchan), R
 
 * Added an example notebook (!46)
 * Added documentation pdf (!46)
+* Added `variogram.SphericalVariogram` (!69)
 * `ellipse`, `ellipse_builder`, and `ellipse_covariance` are renamed to `ellipse.model`,
   `ellipse.estimate`, and `ellipse.covariance` respectively. Ellipse classes `EllipseModel`,
   `EllipseBuilder`, and `EllipseCovarianceBuilder` are available at the `ellipse` level (!66)